+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## [0.1.0] - 13/11/2025
 
 ### Fixed
+- Lowercase in controller registerProduct error.
+- Image saved as a file in register product.
 - Fix the routes to match the frontend
 
 ## [0.1.0] - 11/11/2025
 
+### Fixed
 - Solved the issue that only returns one supply batch changed fixed to send more
 - Fix the search workshops in controllers.
 - Beneficiary querys (update and fetch)